Fix VersionManager Test Code
